DT-1954 Move Postgres initialization into its own file
