+++ src/test/kotlin/uk/gov/justice/digital/hmpps/courtregister/resource/IntegrationTest.kt
@@ -10,39 +10,9 @@ import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.context.DynamicPropertyRegistry
 import org.springframework.test.context.DynamicPropertySource
 import org.springframework.test.web.reactive.server.WebTestClient
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.containers.wait.strategy.Wait
+import uk.gov.justice.digital.hmpps.courtregister.config.PostgresqlContainer
 import uk.gov.justice.digital.hmpps.courtregister.helper.FlywayRestoreExtension
 import uk.gov.justice.digital.hmpps.courtregister.helper.JwtAuthHelper
-import java.io.IOException
-import java.net.ServerSocket
-
-object PostgresqlContainer {
-  val instance: PostgreSQLContainer<Nothing>? by lazy { startPostgresqlContainer() }
-  fun startPostgresqlContainer(): PostgreSQLContainer<Nothing>? =
-    if (checkPostgresRunning().not()) {
-      PostgreSQLContainer<Nothing>("postgres").apply {
-        withEnv("HOSTNAME_EXTERNAL", "localhost")
-        withExposedPorts(5432)
-        withDatabaseName("court_register_db")
-        withUsername("admin")
-        withPassword("admin_password")
-        setWaitStrategy(Wait.forListeningPort())
-        withReuse(true)
-        start()
-      }
-    } else {
-      null
-    }
-
-  private fun checkPostgresRunning(): Boolean =
-    try {
-      val serverSocket = ServerSocket(5432)
-      serverSocket.localPort == 0
-    } catch (e: IOException) {
-      true
-    }
-}
 
 @ExtendWith(FlywayRestoreExtension::class)
 @Suppress("SpringJavaInjectionPointsAutowiringInspection")
